fix: fall back on manifest.licenses (#4) (#5)
* fix: correct plugin import command in README (#4)

* fix: parse all licenses in manifest.licenses (#4)

If the license property is absent from the manifest, and the deprecated
licenses property contains an array of licenses, then fetch all of them
rather than only the first. Conservatively coerce them to a conjunctive
(ANDs) SPDX license expression since it is ambiguous whether the
licenses are meant to be interpreted conjunctively (ANDs) or
disjunctively (ORs).