fix: don't parenthesize single licenses (#10)
If the license (or, more likely, the deprecated licenses property) of a
manifest contains an array of length 1, then don't wrap the SPDX
expression we emit in unnecessary parentheses. For example, while
"(MIT)" is technically a valid SPDX expression, it is a counterintuitive
way of expressing "MIT" and may complicate the implementation of some
downstream config files.